+++ tests/bootstrap.php
@@ -1,4 +1,5 @@
 <?php
 declare(strict_types=1);
 
+require __DIR__ . '/../Simple-PHP-IPAM/version.php';
 require __DIR__ . '/../Simple-PHP-IPAM/lib.php';